requirements.txt
aiofiles
discord.py>=1.6.0
python-dotenv